pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swing-trader"
version = "0.1.0"
description = "News-based swing trading system with multi-signal backtesting and automated execution"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "yfinance>=0.2.30",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "alpaca-trade-api>=3.0",
    "apscheduler>=3.10",
    "ta>=0.11",
    "transformers>=4.35",
    "torch>=2.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "streamlit>=1.28",
    "python-telegram-bot>=20.0",
    "feedparser>=6.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
swing-trader = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
